[Android] Fix casting of Location.getCurrentPositionAsync#timeout param (#2369)

+++ packages/expo-location/android/src/main/java/expo/modules/location/LocationModule.java
@@ -127,7 +127,7 @@ public class LocationModule extends ExportedModule implements ModuleRegistryCons
   @ExpoMethod
   public void getCurrentPositionAsync(final Map<String, Object> options, final Promise promise) {
     // Read options
-    final Long timeout = options.containsKey("timeout") ? (long) options.get("timeout") : null;
+    final Long timeout = options.containsKey("timeout") ? ((Double) options.get("timeout")).longValue() : null;
     boolean highAccuracy = options.containsKey("enableHighAccuracy") && (boolean) options.get("enableHighAccuracy");
 
     final LocationParams locationParams = highAccuracy ? LocationParams.NAVIGATION : LocationParams.BEST_EFFORT;
